+++ lib/qiita/markdown.rb
@@ -8,7 +8,7 @@ require "pygments"
 require "sanitize"
 
 require "qiita/markdown/filters/checkbox"
-require "qiita/markdown/filters/code"
+require "qiita/markdown/filters/code_block"
 require "qiita/markdown/filters/emoji"
 require "qiita/markdown/filters/external_link"
 require "qiita/markdown/filters/final_sanitizer"