+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Rename `Code` to `CodeBlock`
+
 ## 0.20.1
 
 - Fix to sanitize `<input>` which was unexpectedly permitted